feat(TDP-1415): refactored newStyles into styles
