+++ packages/article-magazine-comment/src/article-meta/article-meta.js
@@ -4,6 +4,7 @@ import {
   hasBylineData
 } from "@times-components/article-byline";
 import DatePublication from "@times-components/date-publication";
+import { checkStylesForUnits } from "@times-components/utils";
 
 import metaPropTypes from "./article-meta-prop-types";
 import {
@@ -11,21 +12,21 @@ import {
   Meta,
   MetaContainer,
   Separator
-} from "../newStyles/responsive";
-import newStyles from "../newStyles";
+} from "../styles/responsive";
+import styles from "../styles";
 
 const ArticleMeta = ({ bylines, publicationName, publishedTime }) => (
   <MetaContainer>
     {hasBylineData(bylines) && (
       <Fragment>
-        <Meta styles={newStyles.meta}>
+        <Meta styles={styles.meta}>
           <ArticleBylineWithLinks ast={bylines} />
         </Meta>
         <Separator />
       </Fragment>
     )}
-    <Meta style={newStyles.meta}>
-      <DatePublicationContainer styles={newStyles.datePublication}>
+    <Meta style={styles.meta}>
+      <DatePublicationContainer styles={checkStylesForUnits(styles.datePublication)}>
         <DatePublication date={publishedTime} publication={publicationName} />
       </DatePublicationContainer>
     </Meta>
