+++ packages/article-magazine-comment/src/styles/responsive.js
