+++ packages/article-magazine-comment/src/styles/shared.js
@@ -69,7 +69,8 @@ const sharedStyles = {
     borderRightColor: colours.functional.keyline,
     borderRightWidth: 1,
     height: spacing(3),
-    marginHorizontal: spacing(2)
+    marginLeft: spacing(2),
+    marginRight: spacing(2)
   },
   standFirst: {
     ...fontFactory({
@@ -78,7 +79,8 @@ const sharedStyles = {
     }),
     color: colours.functional.primary,
     marginBottom: spacing(3),
-    paddingHorizontal: spacing(2),
+    paddingLeft: spacing(2),
+    paddingRight: spacing(2),
     textAlign: "center"
   }
 };
